[#110] Feature : 코드 리팩토링
Former-commit-id: e7ce4613ce9b99435c0c3012a82a3ec40a406f90

+++ BE/app/routers/image/permission_router.py
@@ -6,6 +6,7 @@ from typing import List
 from dto.common_dto import CommonResponse
 from configs.mariadb import get_database_mariadb
 from configs.mongodb import get_database_mongodb
+from services.auth.auth_service import JWTManage, Permissions
 from services.image.image_extra_service import ImageExtraService
 from dto.image_detail_dto import ImageUserPermissionRemoveRequest, ImageDepartmentPermissionRemoveRequest, ImageDepartmentPermissionAddRequest, ImageUserPermissionAddRequest
 
@@ -22,8 +23,17 @@ async def add_department_permission(
     mongodb : Session = Depends(get_database_mongodb)
 ):
     try:    
+        access_token = credentials.credentials
+        jwt = JWTManage(maria_db)
+        user_id = jwt.verify_token(access_token)["user_id"]
+        
+        permission = Permissions(maria_db, mongodb)
+        ids = await permission.get_image_permissions(user_id)
+        
+        if request.image_id not in ids:
+            raise HTTPException(status_code=403, detail="Permission Denied")
+        
         image_extra_service = ImageExtraService(maria_db, mongodb)
-
         response = await image_extra_service.add_department_image_permission(request)
 
         return CommonResponse(
@@ -44,6 +54,16 @@ async def add_user_permission(
     mongodb : Session = Depends(get_database_mongodb)
 ):
     try:    
+        access_token = credentials.credentials
+        jwt = JWTManage(maria_db)
+        user_id = jwt.verify_token(access_token)["user_id"]
+        
+        permission = Permissions(maria_db, mongodb)
+        ids = await permission.get_image_permissions(user_id)
+        
+        if request.image_id not in ids:
+            raise HTTPException(status_code=403, detail="Permission Denied")
+        
         image_extra_service = ImageExtraService(maria_db, mongodb)
 
         response = await image_extra_service.add_user_image_permission(request)
@@ -66,6 +86,16 @@ async def remove_user_permission(
     mongodb : Session = Depends(get_database_mongodb)
 ):
     try:
+        access_token = credentials.credentials
+        jwt = JWTManage(maria_db)
+        user_id = jwt.verify_token(access_token)["user_id"]
+        
+        permission = Permissions(maria_db, mongodb)
+        ids = await permission.get_image_permissions(user_id)
+        
+        if request.image_id not in ids:
+            raise HTTPException(status_code=403, detail="Permission Denied")
+        
         image_extra_service = ImageExtraService(maria_db, mongodb)
 
         response = await image_extra_service.remove_user_image_permission(request)
@@ -88,6 +118,16 @@ async def remove_department_permission(
     mongodb : Session = Depends(get_database_mongodb)
 ):
     try:
+        access_token = credentials.credentials
+        jwt = JWTManage(maria_db)
+        user_id = jwt.verify_token(access_token)["user_id"]
+        
+        permission = Permissions(maria_db, mongodb)
+        ids = await permission.get_image_permissions(user_id)
+        
+        if request.image_id not in ids:
+            raise HTTPException(status_code=403, detail="Permission Denied")
+        
         image_extra_service = ImageExtraService(maria_db, mongodb)
 
         response = await image_extra_service.remove_department_image_permission(request)
